Isolate External Interface to nes.dll in one place
And remove all platforms except for x86
